feat(wds-codemod): add codemod for dialog button migration
